Custom capped atomic map PoC
Add a custom implementation of a concurrent-safe map based on the old Go
`sync.Map` implementation. This type, `cappedMap`, is a reduced API that
is limited to a maximum capacity. Any addition beyond that capacity is
added with the overflow attribute set.

This custom type is used in conjunction with any set attribute limits.
If not attribute limits are set, the `sync.Map` is used.

This is a proof-of-concept. Current tests pass, but it is missing
coverage of the added functionality and integration. More work is needed
before this can considered ready for integration with the code base.